Move `approx_curve` to `approx::curve`
